Fix array subscript has type 'char'.

+++ src/kernel.c
@@ -484,7 +484,7 @@ char *cmdtok(char *cmd)
 					quo = *end;
 				*end = '\0';
 			}
-			else if (isspace(*end) && !quo)
+			else if (isspace((int)*end) && !quo)
 				*end = '\0';
 		}
 	}
@@ -550,7 +550,7 @@ void find_events()
 	for (; *p; p++) {
 		if (*p == '!') {
 			q = p;
-			while (*q && !isspace(*q))
+			while (*q && !isspace((int)*q))
 				q++;
 			for (i = cur_his + HISTORY_COUNT - 1; i > cur_his; i--) {
 				if (!strncmp(cmd[i % HISTORY_COUNT], p + 1, q - p - 1)) {
@@ -585,7 +585,7 @@ int fill_arg(char *const dest, const char *argv)
 	char *p = NULL;
 
 	for (; *argv; argv++) {
-		if (isalnum(*argv) || *argv == '_') {
+		if (isalnum((int)*argv) || *argv == '_') {
 			if (p)
 				*p++ = *argv;
 			else
